[NotificationManager/AN-1628] - notifications not saved on db Notifications were not being inserted in database. In order to obtain notifications the following steps are made: -obtain an account (observable) -make the request (observable) -save notifications on database (single)
when we obtain the account, we obtain an observable that never completes and after the request we were changing that observable into a single. Since the account observable never completes, the single never emits, if single never emits, notifications won't be saved on database.

the problem was solved by getting only the first account emitted(now the observable completes!).