Update settings.gradle
The original code lacked the ".php" extension in the "include" line. This can cause the code to not work properly because the file cannot be found. In the corrected code, the ".php" extension is appended to each "include" line to ensure that the correct files are included. In addition, the code has been reformatted and formatted to improve readability.