[DT-1311] Encapsulate payment methods logic
Currently this logic is split between either PaymentMethod and ViewModel
or between PaymentMethod and *Transaction. This complicates ViewModels
and transaction logic. In order to simplify it's better to encapsulate
the logic of additional step into the payment method itself. And do the
same with the UI logic, consolidating it in UI states.
This also helps to achieve the goal of the task since the payment method
specific transactions lack necessary information to return with results.
And to get that information we have to call a generic transactions API.