[DT-1311] Extract transaction to make it full and independent
Full in terms of necessary data returned from server and independent
from the specific payment methods transaction creation variations.
Here is the problem. To pass back result the purchase we need to get it
from products repository by purchase uid. And from architecture
perspective it makes sense to get that from the Transaction itself.
2 problems arise here:
1. The payment methods gateways don't have purchase uid in the response.
So we need to use a payment method independent transaction API endpoint
to get it (we already use it to poll for the transaction state).
2. In order to get the purchase uid by a specific payment method
transaction it needs to know (be coupled) with the products module.
Taking this into account it makes sense to decouple the transaction from
a specific payment methods. Both to avoid coupling payment methods with
the products module and to not repeat the transaction polling logic for
every payment method we have or will have in the future.